requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23

# PDF Generation
reportlab==4.0.8

# Barcode Generation
python-barcode==0.15.1
Pillow==10.1.0

# Date handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Xero Integration
xero-python==4.3.0
PyJWT==2.8.0
requests-oauthlib==1.3.1

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
